test/requirements.txt
azure-servicebus==7.0.0
azure-identity==1.5.0
bpython==0.20.1